requirements.txt
colpali-engine
torch
Pillow
pymupdf
qdrant-client
openai
streamlit
pydantic-settings
